[Worker] Moves each consumer bootstep into separated module.
