+++ celery/worker/consumer/__init__.py
@@ -1,0 +1,17 @@
+from __future__ import absolute_import, unicode_literals
+
+from .consumer import Consumer
+
+from .agent import Agent
+from .connection import Connection
+from .control import Control
+from .events import Events
+from .gossip import Gossip
+from .heart import Heart
+from .mingle import Mingle
+from .tasks import Tasks
+
+__all__ = [
+    'Consumer', 'Agent', 'Connection', 'Control',
+    'Events', 'Gossip', 'Heart', 'Mingle', 'Tasks',
+]
